Push isize through fdtable, pipe, and IPC layers
Complete the native-width conversion through the last remaining APIs:

- rux-ipc pipe: alloc/read/write return i64 → isize
- rux-vfs fdtable: all function returns i64 → isize, PipeOps trait → isize
- kernel fdtable wrapper: all returns i64 → isize
- kernel pipe wrapper: PipeOps impl + create() → isize
- Remove redundant `as isize` casts in posix.rs/linux.rs now that
  underlying APIs return isize directly

lseek keeps i64 internally for file offset arithmetic (files can
exceed isize range on 32-bit), casts to isize only on return.

+++ crates/kernel/src/fdtable.rs
@@ -12,27 +12,27 @@ pub use rux_vfs::fdtable::{
 use crate::pipe::PIPE_OPS;
 
 /// Open a file by path. Supplies kernel VFS automatically.
-pub fn sys_open(path: &[u8]) -> i64 {
+pub fn sys_open(path: &[u8]) -> isize {
     unsafe { rux_vfs::fdtable::sys_open(path, crate::kstate::fs()) }
 }
 
 /// Open a file by inode with flags.
-pub fn sys_open_ino(ino: rux_vfs::InodeId, flags: u32) -> i64 {
+pub fn sys_open_ino(ino: rux_vfs::InodeId, flags: u32) -> isize {
     unsafe { rux_vfs::fdtable::sys_open_ino(ino, flags, crate::kstate::fs()) }
 }
 
 /// Duplicate a file descriptor.
-pub fn sys_dup(oldfd: usize) -> i64 {
+pub fn sys_dup(oldfd: usize) -> isize {
     rux_vfs::fdtable::sys_dup(oldfd)
 }
 
 /// Duplicate a file descriptor to the lowest fd >= minfd.
-pub fn sys_dupfd(oldfd: usize, minfd: usize) -> i64 {
+pub fn sys_dupfd(oldfd: usize, minfd: usize) -> isize {
     rux_vfs::fdtable::sys_dupfd(oldfd, minfd)
 }
 
 /// Duplicate fd to a specific newfd.
-pub fn sys_dup2(oldfd: usize, newfd: usize) -> i64 {
+pub fn sys_dup2(oldfd: usize, newfd: usize) -> isize {
     unsafe {
         rux_vfs::fdtable::sys_dup2(
             oldfd, newfd,
@@ -43,7 +43,7 @@ pub fn sys_dup2(oldfd: usize, newfd: usize) -> i64 {
 }
 
 /// Close a file descriptor.
-pub fn sys_close(fd: usize) -> i64 {
+pub fn sys_close(fd: usize) -> isize {
     unsafe {
         rux_vfs::fdtable::sys_close(
             fd,
@@ -54,7 +54,7 @@ pub fn sys_close(fd: usize) -> i64 {
 }
 
 /// Read from a file descriptor.
-pub fn sys_read_fd(fd: usize, buf: *mut u8, len: usize) -> i64 {
+pub fn sys_read_fd(fd: usize, buf: *mut u8, len: usize) -> isize {
     unsafe {
         rux_vfs::fdtable::sys_read_fd(
             fd, buf, len,
@@ -65,7 +65,7 @@ pub fn sys_read_fd(fd: usize, buf: *mut u8, len: usize) -> i64 {
 }
 
 /// Write to a file descriptor.
-pub fn sys_write_fd(fd: usize, buf: *const u8, len: usize) -> i64 {
+pub fn sys_write_fd(fd: usize, buf: *const u8, len: usize) -> isize {
     unsafe {
         rux_vfs::fdtable::sys_write_fd(
             fd, buf, len,
@@ -76,7 +76,7 @@ pub fn sys_write_fd(fd: usize, buf: *const u8, len: usize) -> i64 {
 }
 
 /// Seek on a file descriptor.
-pub fn sys_lseek(fd: usize, offset: i64, whence: u32) -> i64 {
+pub fn sys_lseek(fd: usize, offset: i64, whence: u32) -> isize {
     unsafe {
         rux_vfs::fdtable::sys_lseek(fd, offset, whence, crate::kstate::fs())
     }
